refactor parse_allfilters_c
